pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sera"
version = "0.1.0"
description = "Self-Evolving Research Agent"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "accelerate>=1.0.0",
    "safetensors>=0.4.0",
    "typer>=0.12.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "graphviz>=0.20.0",
    "numpy>=1.26.0",
    "pydantic>=2.6.0",
    "rich>=13.7.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-mock>=3.12", "respx>=0.21", "ruff>=0.3"]
slurm = ["submitit>=1.5.0"]
docker = ["docker>=7.0.0"]
vllm = ["vllm>=0.15.0"]
latex = ["pandoc>=2.3"]

[project.scripts]
sera = "sera.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/sera"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
